Bumper Car problem - Operating Systems Summer 2020
For this problem, we were asked to use threads to implement a bumper car park; a rider should be able to ride a bumper car for a variable amount of time, then walk around the park before waiting in line to ride again. The amount of riders, bumper cars, and the amount of time you would want to run the simulation is all gotten from user inputs.